Add in a database to Ariel
Defaults to an in-memory SQLite database, but can
be either MySQL or SQLite.

Add in migrations to add a `songs` table and an `ao3_urls` table.

The `songs` table will replace the current json file usage, while the
`ao3_urls` table will handle the URLs for songs, and any other instances